redesign landing, not responsive yet
