+++ 404.html
@@ -1,1 +1,1 @@
-<html><head><meta charset="utf-8"/><title>owo &middot; 404</title><meta http-equiv="X-UA-Compatible" content="IE=edge"/><meta name="viewport" content="width=device-width, initial-scale=1"/><link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/meyer-reset/2.0/reset.min.css"/><link rel="stylesheet" href="/assets/css/main.css"/><link rel="icon" type="image/png" href="assets/favicon.png"/></head><body><main><header><h1>OwO, whats th-</h1><h3>Oh, no bulge here...</h3></header><section><p>We can't find that page, maybe return <a href="index.html">home</a>?</p></section><footer><ul><li><a href="index.html">owo</a></li><li><a href="faq.html">faq</a></li><li><a href="https://github.com/whats-this/transparency">transparency</a></li><li><a href="https://github.com/whats-this">github</a></li><li><a href="mailto:dean@whats-th.is">contact</a></li></ul></footer></main></body></html>
+<!DOCTYPE html><html><head><meta charset="utf-8"><title>owo &middot; 404</title><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width, initial-scale=1"><link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:400,700"><link rel="stylesheet" href="/assets/css/styles.css"><link rel="icon" type="image/png" href="/assets/img/favicon.png"></head><body><div class="content-box"><div class="page-content"><header><h1><a href="/">OWO</a></h1><h2>404 Not Found</h2></header><p>We can't find that page, maybe try returning <a href="/">home</a>?</p></div><footer><ul><li><a href="/">HOME</a></li><li><a href="/faq.html">FAQ</a></li><li><a href="https://github.com/whats-this/transparency">TRANSPARENCY</a></li><li><a href="https://github.com/whats-this">GITHUB</a></li><li><a href="&#109;ailto&#58;&#100;ean&#64;&#37;77ha&#116;&#115;&#45;%74h&#46;is">CONTACT</a></li></ul></footer></div></body></html>
