+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2016 OwO.Whats-Th.is?
+Copyright (c) 2016 Dean Sheather
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
