+++ assets/css/styles.css
@@ -1,0 +1,1 @@
+*{box-sizing:border-box}/*! normalize.css v5.0.0 | MIT License | github.com/necolas/normalize.css */html{font-family:sans-serif;line-height:1.15;-ms-text-size-adjust:100%;-webkit-text-size-adjust:100%}body{margin:0}article,aside,footer,header,nav,section{display:block}h1{font-size:2em;margin:0.67em 0}figcaption,figure,main{display:block}figure{margin:1em 40px}hr{box-sizing:content-box;height:0;overflow:visible}pre{font-family:monospace, monospace;font-size:1em}a{background-color:transparent;-webkit-text-decoration-skip:objects}a:active,a:hover{outline-width:0}abbr[title]{border-bottom:none;text-decoration:underline;text-decoration:underline dotted}b,strong{font-weight:inherit}b,strong{font-weight:bolder}code,kbd,samp{font-family:monospace, monospace;font-size:1em}dfn{font-style:italic}mark{background-color:#ff0;color:#000}small{font-size:80%}sub,sup{font-size:75%;line-height:0;position:relative;vertical-align:baseline}sub{bottom:-0.25em}sup{top:-0.5em}audio,video{display:inline-block}audio:not([controls]){display:none;height:0}img{border-style:none}svg:not(:root){overflow:hidden}button,input,optgroup,select,textarea{font-family:sans-serif;font-size:100%;line-height:1.15;margin:0}button,input{overflow:visible}button,select{text-transform:none}button,html [type="button"],[type="reset"],[type="submit"]{-webkit-appearance:button}button::-moz-focus-inner,[type="button"]::-moz-focus-inner,[type="reset"]::-moz-focus-inner,[type="submit"]::-moz-focus-inner{border-style:none;padding:0}button:-moz-focusring,[type="button"]:-moz-focusring,[type="reset"]:-moz-focusring,[type="submit"]:-moz-focusring{outline:1px dotted ButtonText}fieldset{border:1px solid #c0c0c0;margin:0 2px;padding:0.35em 0.625em 0.75em}legend{box-sizing:border-box;color:inherit;display:table;max-width:100%;padding:0;white-space:normal}progress{display:inline-block;vertical-align:baseline}textarea{overflow:auto}[type="checkbox"],[type="radio"]{box-sizing:border-box;padding:0}[type="number"]::-webkit-inner-spin-button,[type="number"]::-webkit-outer-spin-button{height:auto}[type="search"]{-webkit-appearance:textfield;outline-offset:-2px}[type="search"]::-webkit-search-cancel-button,[type="search"]::-webkit-search-decoration{-webkit-appearance:none}::-webkit-file-upload-button{-webkit-appearance:button;font:inherit}details,menu{display:block}summary{display:list-item}canvas{display:inline-block}template{display:none}[hidden]{display:none}html,body{height:100%}body{background-color:#252525;background-image:url(../img/background-tile.jpg)}.content-box{height:100%;padding:200px 100px 0 100px;display:flex;flex-direction:column}@media only screen and (max-width: 1400px){.content-box{padding-top:100px}}@media only screen and (max-width: 900px){.content-box{padding-top:50px}}.content-box>.page-content{flex:1 0 auto}.content-box>.page-content::after{height:0px;content:'\00a0';display:block;visibility:hidden}.content-box>footer{margin-top:40px;flex:none}.faq-box{width:50%}.faq-box>h3{margin:30px 0 0 0;font-weight:bold;background-color:#fff;color:#252525;display:inline-block;border-radius:5px}.faq-box>h3>a{padding:4px 13px;display:inline-block;position:relative}.faq-box>h3>a:hover:before{display:block}.faq-box>h3>a:before{padding-right:13px;position:absolute;top:5px;left:-18px;display:none;color:#fff;content:"#"}.faq-box>h4{margin-bottom:10px;font-weight:bold}.faq-box>ul{margin:0}footer{margin-bottom:30px}footer ul{padding:0}footer ul li:not(:first-child){margin-left:20px}footer ul li{list-style:none;display:inline-block}footer ul li a{font-size:1.1rem;text-decoration:none;line-height:1.1rem}footer ul li a:hover{color:#448CCB}header{max-width:660px;margin-bottom:50px;line-height:1}header h1{margin:0;color:#448CCB;font-weight:bold;font-size:130px}header h1>a{text-decoration:none}header h2{margin:10px 0 0 0;font-size:52px}input{padding:6px 13px;background:none;border:none;outline:0;color:inherit;border-left:3px solid #fff;border-top-left-radius:2px;border-bottom-left-radius:2px}input:hover{cursor:text}input::-webkit-input-placeholder{color:inherit}input:-moz-placeholder{color:inherit}input::-moz-placeholder{color:inherit}input:-ms-input-placeholder{color:inherit}a{color:inherit;text-decoration:underline}a:active,a:hover,a:visited{color:inherit}body{font-family:Roboto,sans-serif;font-weight:400;font-size:16px;line-height:1.4;color:#fff}h2,h3,h4,h5,h6{font-weight:normal}
