+++ index.html
@@ -1,1 +1,18 @@
-<html><head><meta charset="utf-8"/><title>owo &middot; home</title><meta http-equiv="X-UA-Compatible" content="IE=edge"/><meta name="viewport" content="width=device-width, initial-scale=1"/><link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/meyer-reset/2.0/reset.min.css"/><link rel="stylesheet" href="/assets/css/main.css"/><link rel="icon" type="image/png" href="assets/favicon.png"/></head><body><main><header><h1 id="owo">owo</h1><h3>Private pomf-compatible file host<br/>& Private polr URL shortener</h3></header><section><p>All accounts are given with approval from Dean.</p><input id="email" value="dean@whats-th.is" readonly="readonly"/></section><footer><ul><li><a href="index.html">owo</a></li><li><a href="faq.html">faq</a></li><li><a href="https://github.com/whats-this/transparency">transparency</a></li><li><a href="https://github.com/whats-this">github</a></li><li><a href="mailto:dean@whats-th.is">contact</a></li></ul></footer></main></body><script type="text/javascript" src="assets/script.js"></script></html>
+<!DOCTYPE html><html><head><meta charset="utf-8"><title>owo &middot; home</title><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width, initial-scale=1"><link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:400,700"><link rel="stylesheet" href="/assets/css/styles.css"><link rel="icon" type="image/png" href="/assets/img/favicon.png"></head><body><div class="content-box"><div class="page-content"><header><h1><a href="/">OWO</a></h1><h2>A private, Pomf-compatible file host.</h2></header><p>All accounts are given with approval from Dean
+<small>(<a href="https://keybase.io/deansheather" target="_blank">Keybase</a>)</small>:</p><input id="email" value="dean@whats-th.is" readonly></div><footer><ul><li><a href="/">HOME</a></li><li><a href="/faq.html">FAQ</a></li><li><a href="https://github.com/whats-this/transparency">TRANSPARENCY</a></li><li><a href="https://github.com/whats-this">GITHUB</a></li><li><a href="&#109;ailto&#58;&#100;ean&#64;&#37;77ha&#116;&#115;&#45;%74h&#46;is">CONTACT</a></li></ul></footer></div><script type="text/javascript">var email = document.getElementById('email');
+
+email.onclick = email.onfocus = function () {
+  if (this.__active) return;
+  this.__active = true;
+  var self = this;
+
+  setTimeout(function () {
+    if (self.setSelectionRange) self.setSelectionRange(0, self.value.length);
+    else self.select();
+  }, 0);
+};
+
+email.onblur = function () {
+  this.__active = false;
+};
+</script></body></html>
