+++ public-cdn-domains.txt
@@ -1,0 +1,18 @@
+files:owo.whats-th.is
+files:all.foxgirlsare.sexy
+files:buttsare.sexy
+files:can-i-ask-dean-on-a.date
+files:catch-me-outside.how-about-th.at
+files:deanis.sexy
+files:how-about-th.at
+files:i.am-a.ninja
+files:i.stole-a-me.me
+files:im.not.ovh
+files:light-theme-hurts.my-ey.es
+files:my.catgirl.party
+files:nyanyanya.moe
+files:this.is-a.dog
+files:totally-not.a-sketchy.site
+
+link:awau.moe
+link:uwu.whats-th.is
