--- src/scss/main.scss
@@ -1,143 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto:400,700,900');
-
-$main-color: #448CCB;
-$font-color: #fff;
-$darker-color: #ccc;
-
-body, html {
-    font-family:Roboto, sans-serif;
-    background: $main-color;
-    color: $font-color;
-    width:100%;
-}
-main {
-    display:flex;
-    flex-direction:column;
-    align-items:center;
-    justify-content:center;
-    text-align:center;
-}
-header {
-    margin-top:2rem;
-    h1 {
-        font-weight:bold;
-        font-size:8rem;
-        padding-top:2%;
-        cursor:pointer;
-    }
-    h3 {
-        font-size:1rem;
-        font-weight:bold;
-        text-transform:uppercase;
-        margin:.8rem 0 0 0;
-    }
-}
-section {
-    margin-top:2rem;
-    width:100%;
-    max-width:540px;
-    p {
-        a {
-            color:$font-color;
-            text-decoration:none;
-            font-weight:bold;
-            &:hover {
-                color:$darker-color;
-            }
-        }
-    }
-    input {
-        margin-top:1.5rem;
-        padding:.6rem 5rem;
-        background:$font-color;
-        color:$main-color;
-        font-weight:bold;
-        border-radius:2px;
-        border:none;
-        text-align:center;
-        font-size:1rem;
-    }
-    h4 {
-        padding:0.6rem 1.25rem;
-        background:$font-color;
-        color:$main-color;
-        text-align:left;
-        text-transform:uppercase;
-        font-weight:bold;
-        border-radius:2px;
-    }
-    #a {
-        padding:0.6rem 1.25rem 2.5rem 1.25rem;
-        line-height:1.25rem;
-        text-align:left;
-        a {
-            text-decoration:none;
-            color:$font-color;
-            font-weight:bold;
-            &:hover {
-                color:$darker-color;
-            }
-        }
-    }
-    h5#a {
-        padding:1rem;
-        text-align:center;
-    }
-    h6#a {
-        font-weight:bold;
-        padding:0.6rem;
-        font-size:0.75rem;
-        text-transform:uppercase;
-        text-align:center;
-    }
-    ul {
-        a {
-            text-decoration:none;
-            li {
-                color:$font-color;
-
-            }
-        }
-        padding:0.6rem 1.25rem 2.5rem 1.25rem;
-        li {
-            padding:.9rem .3rem;
-            background:rgba(255,255,255,0.1);
-            margin:.3rem;
-            text-align:center;
-            border-radius:2px;
-            a {
-                color:$font-color;
-                text-decoration:none;
-                cursor:pointer;
-                &:hover {
-                    color:$darker-color;
-                }
-            }
-        }
-    }
-}
-footer {
-    margin-top:2.5rem;
-    margin-bottom:2.5rem;
-    width:300px;
-
-    ul {
-        min-width: 100vw;
-        margin-left:-22px;
-        li:not(:first-child)::before {
-            content:'・';
-        }
-        li {
-            float:left;
-            a {
-                color:$font-color;
-                font-size:1.1rem;
-                text-decoration:none;
-                line-height:1.1rem;
-                &:hover {
-                    color:$darker-color;
-                }
-            }
-        }
-    }
-}
