+++ src/scss/modules/_media.scss
@@ -1,0 +1,9 @@
+@mixin media($media) {
+  @if $media == sm {
+    @media only screen and (max-width: $sm) { @content; }
+  } @else if $media == md {
+    @media only screen and (max-width: $md) { @content; }
+  } @else if $media == lg {
+    @media only screen and (max-width: $lg) { @content; }
+  }
+}
