+++ src/scss/partials/_background.scss
@@ -1,0 +1,8 @@
+html, body {
+  height: 100%;
+}
+
+body {
+  background-color: $background-color;
+  background-image: url(../img/background-tile.jpg);
+}
