+++ src/scss/partials/_footer.scss
@@ -1,0 +1,22 @@
+footer {
+  margin-bottom: 30px;
+
+  ul {
+    padding: 0;
+    li:not(:first-child) {
+      margin-left: 20px;
+    }
+    li {
+      list-style: none;
+      display: inline-block;
+      a {
+        font-size:1.1rem;
+        text-decoration:none;
+        line-height:1.1rem;
+        &:hover {
+            color: $brand-primary;
+        }
+      }
+    }
+  }
+}
