+++ src/scss/partials/_header.scss
@@ -1,0 +1,19 @@
+header {
+  max-width: 660px;
+  margin-bottom: 50px;
+  line-height: 1;
+
+  h1 {
+    margin: 0;
+    color: $brand-primary;
+    font-weight: bold;
+    font-size: 130px;
+
+    > a { text-decoration: none; }
+  }
+
+  h2 {
+    margin: 10px 0 0 0;
+    font-size: 52px;
+  }
+}
