+++ src/scss/partials/_inputs.scss
@@ -1,0 +1,17 @@
+input {
+  padding: 6px 13px;
+  background: none;
+  border: none;
+  outline: 0;
+  color: inherit;
+  border-left: 3px solid $text-color;
+  border-top-left-radius: 2px;
+  border-bottom-left-radius: 2px;
+
+  &:hover { cursor: text; }
+
+  &::-webkit-input-placeholder { color: inherit; }
+  &:-moz-placeholder { color: inherit; }
+  &::-moz-placeholder { color: inherit; }
+  &:-ms-input-placeholder { color: inherit; }
+}
