+++ src/scss/partials/_links.scss
@@ -1,0 +1,6 @@
+a {
+  color: inherit;
+  text-decoration: underline;
+
+  &:active, &:hover, &:visited { color: inherit; }
+}
