+++ src/scss/styles.scss
@@ -1,0 +1,13 @@
+// Base
+@import "partials/base";
+
+// Partials
+@import "partials/normalize";
+@import "partials/background";
+@import "partials/content-box";
+@import "partials/faq";
+@import "partials/footer";
+@import "partials/header";
+@import "partials/inputs";
+@import "partials/links";
+@import "partials/typography";
