+++ src/views/404.pug
@@ -1,28 +1,8 @@
-html
-  head
-    meta(charset='utf-8')
-    title owo &middot; 404
-    meta(http-equiv='X-UA-Compatible', content='IE=edge')
-    meta(name='viewport', content='width=device-width, initial-scale=1')
-    link(rel='stylesheet', href='https://cdnjs.cloudflare.com/ajax/libs/meyer-reset/2.0/reset.min.css')
-    link(rel='stylesheet', href='/assets/css/main.css')
-    link(rel='icon', type='image/png', href='assets/favicon.png')
-  body
-    main
-      header
-       h1 OwO, whats th-
-       h3 Oh, no bulge here...
-      section
-        p We can't find that page, maybe return #[a(href='index.html') home]?
-      footer
-        ul
-          li
-            a(href='index.html') owo
-          li
-            a(href='faq.html') faq
-          li
-            a(href='https://github.com/whats-this/transparency') transparency
-          li
-            a(href='https://github.com/whats-this') github
-          li
-            a(href='mailto:dean@whats-th.is') contact
+extends layouts/_owo.pug
+
+block variables
+  - var htmlTitle = "404";
+  - var pageTitle = "404 Not Found";
+
+block content
+  p We can't find that page, maybe try returning #[a(href='/') home]?
