+++ src/views/partials/_footer.pug
@@ -1,0 +1,14 @@
+footer
+  ul
+    li
+      a(href='/') HOME
+    li
+      a(href='/faq.html') FAQ
+    li
+      a(href='https://github.com/whats-this/transparency') TRANSPARENCY
+    li
+      a(href='https://github.com/whats-this') GITHUB
+    //- li
+    //-   a(href='https://whats-th.is/donate') DONATE
+    li
+      <a href="&#109;ailto&#58;&#100;ean&#64;&#37;77ha&#116;&#115;&#45;%74h&#46;is">CONTACT</a>
