+++ src/views/partials/_header.pug
@@ -1,0 +1,3 @@
+header
+  h1 #[a(href='/') OWO]
+  h2 #{pageTitle}
